% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe Function}
\usage{
describe(var)
}
\arguments{
\item{var}{The only parameter. Must be a quantitative variable.}
}
\value{

}
\description{
Takes a single parameter, a quantitative variable, and returns basic
descriptive plots, including the histogram, density histogram with a density line, a boxplot,
and a quantile plot.
}
\examples{
describe(variable)
}
