% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DID.R
\name{DID}
\alias{DID}
\title{Differences in Differences (DID) Function}
\usage{
DID(dv, time, yr, treat, h)
}
\arguments{
\item{dv}{dependent variable}

\item{time}{the variable that will be converted to time dummy variable in the function}

\item{yr}{the condition for the time dummy variable to be created. Should be a year
or time period}

\item{treat}{the group that is being treated. This is a variable that will be converted to a
dummy variable in the function}

\item{h}{the condition for the treatment dummy variable to be created}
}
\value{

}
\description{
This function is meant to conduct the process for running a
Difference in Differences regression model. It takes 5 parameters and will return the summary
of the DID regression.
}
\examples{
DID(dependent var, timeVariable, yr, treat, h)
}
