% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOM_Parameters.R
\name{MOM_Parameters}
\alias{MOM_Parameters}
\title{Method of Moments Parameters}
\usage{
MOM_Parameters(Var)
}
\arguments{
\item{Var}{}
}
\value{

}
\description{
This function takes a variable and computes its method of moments
parameters, alpha and beta. This is useful for bootstrapping. The funciton accepts
a single parameter
}
\examples{
MOM_Parameters(variable)
}
