% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoccerData.R
\docType{data}
\name{SoccerData}
\alias{SoccerData}
\title{Soccer Data from sportsdata.io}
\format{
6 x 730

\describe{
\item{index}{Name}
\item{column 1}{Team}
\item{column 2}{Minutes}
\item{column 3}{FantasyPoints}
\item{column 4}{Goals}
\item{column 5}{Shots}
}
}
\usage{
SoccerData
}
\description{
Soccer Data from sportsdata.io
}
\keyword{datasets}
